# openapi.yaml
...
openapi: 3.0.3
info:
  title: Panuval Bookstore API
  version: 1.0.0
  description: API specification for the Panuval Tamil Bookstore mobile application.
servers:
  - url: https://api.panuval.com/v1
    description: Production Server
  - url: https://staging-api.panuval.com/v1
    description: Staging Server
paths:
...
  /items/search:
    get:
      summary: Search and Filter Items
      description: Retrieves a list of items based on keyword search and various filter criteria. Supports sorting and pagination.
      parameters:
        - $ref: '#/components/parameters/SearchQueryParam'
        - $ref: '#/components/parameters/CategoryFilterParam'
        - $ref: '#/components/parameters/AuthorFilterParam'
        - $ref: '#/components/parameters/PublisherFilterParam'
        - $ref: '#/components/parameters/PriceMinFilterParam'
        - $ref: '#/components/parameters/PriceMaxFilterParam'
        - $ref: '#/components/parameters/YearFilterParam'
        - $ref: '#/components/parameters/StockFilterParam'
        - $ref: '#/components/parameters/LanguageFilterParam'
        - $ref: '#/components/parameters/ItemTypeFilterParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Successfully retrieved search results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemSearchResponse'
        '400': # For invalid filter/sort values
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /items/{itemId}:
    get:
      summary: Get Item Details
      description: Retrieves detailed information about a specific item using its ID.
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
          description: The unique ID of the item.
      responses:
        '200':
          description: Successfully retrieved item details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: A human-readable error message.
        errorCode:
          type: string
          description: An optional machine-readable error code.
          nullable: true
      required:
        - message
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: General validation error message (e.g., "Validation Failed").
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: The field that failed validation.
              message:
                type: string
                description: The specific validation error message for the field.
            required:
              - field
              - message
      required:
        - message
        - errors
...
    ItemSummary:
      type: object
      properties:
        itemId:
          type: integer
        name:
          type: string
          description: Name of the item.
        imageUrl:
          type: string
          format: url
          nullable: true
        price:
          type: number
          format: float
          description: Current selling price.
        originalPrice:
          type: number
          format: float
          description: Original price, MRP.
        discountPercentage:
          type: number
          format: float
          description: Percentage discount applied. Null if no discount.
          nullable: true
        stockStatus:
          $ref: '#/components/schemas/StockStatus'
        label:
          $ref: '#/components/schemas/ItemLabel'
      required:
        - itemId
        - title
        - price
        - stockStatus
    PaginationInfo:
      type: object
      properties:
        currentPage:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
      required:
        - currentPage
        - pageSize
        - totalItems
        - totalPages
    ItemSearchResponse:
      type: object
      properties:
        displayText:
          type: string
          description: Text to display for the search results.
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemSummary'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
        # Optionally include available facets/filters based on the results
        # availableFilters: { type: object, ... }
      required:
        - items
        - pagination
    StockStatus:
      type: string
      enum: [AVAILABLE, IN_STOCK, OUT_OF_STOCK, PRE_ORDER, OUT_OF_PRINT]
      description: Current stock status of the item
    ItemDetail:
      type: object
      properties:
        itemId:
          type: integer
        title:
          type: string
          description: Name of the item.
        subTitle:
          type: string
          description: Subtitle of the item.
        description:
          type: string
          description: Description of the item.
        coverImageUrl:
          type: string
          format: url
          nullable: true
        moreImages:
          type: array
          items:
            type: string
            format: url
          nullable: true
        price:
          type: number
          format: float
          description: Current selling price.
        originalPrice:
          type: number
          format: float
          description: Discounted price, if applicable. Null otherwise.
          nullable: true
        discountPercentage:
          type: number
          format: float
          description: Percentage discount applied. Null if no discount.
          nullable: true
        stockStatus:
          $ref: '#/components/schemas/StockStatus'
        shortDescription:
          type: string
          description: Short description of the item.
        details:
          type: object
          additionalProperties:
            type: string
          description: Map containing additional details like ISBN, format, binding, etc. about the item.
          nullable: true
        authors:
          type: array
          items:
            $ref: '#/components/schemas/AuthorInfo'
        publishers:
          type: array
          items:
            $ref: '#/components/schemas/PublisherInfo'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryInfo'
        highlights:
          type: array
          items:
            type: string
          description: List of highlights or key features of the item.
        policyText:
          type: string
          description: Policy text related to the item, such as shipping, refund, return policy.
        label:
          $ref: '#/components/schemas/ItemLabel'
      required:
        - itemId
        - title
        - price
        - stockStatus
    ItemLabel:
      type: object
      properties:
        label:
          type: string
          description: Label for the item, such as "Best Seller", "New Arrival", etc.
        showLabel:
          type: boolean
          description: Indicates if the label should be displayed.
          default: true 
        labelType:
          type: string
          enum: [BEST_SELLER, NEW_ARRIVAL, AWARD_WINNING, NONE]
          description: Type of label to be displayed.
    SearchFilter:
      type: string
      description: A string representing the search filter parameters.
      example: 'authorId=123&categoryId=456'
    AuthorInfo:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the publisher
        name:
          type: string
          description: Name of the author, publisher, or category.
        searchFilter:
          $ref: '#/components/schemas/SearchFilter'
        type:
          type: string
          enum: [AUTHOR, TRANSLATOR, EDITOR, ILLUSTRATOR, CONTRIBUTOR]
          description: Type of the author
      required:
        - name
        - searchFilter
    PublisherInfo:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the publisher.
        name:
          type: string
          description: Name of the publisher
        searchFilter:
          $ref: '#/components/schemas/SearchFilter'
      required:
        - name
        - searchFilter
    CategoryInfo:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the category.
        name:
          type: string
          description: Name of the category
        searchFilter:
          $ref: '#/components/schemas/SearchFilter'
      required:
        - name
        - searchFilter
  responses:
    BadRequest:
      description: Bad Request - Often due to validation errors or malformed input.
      content:
        application/json:
          schema:
            oneOf: # Can be a general error or a specific validation error
              - $ref: '#/components/schemas/ErrorResponse'
              - $ref: '#/components/schemas/ValidationErrorResponse'
...
    NotFound:
      description: Not Found - The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal Server Error - An unexpected error occurred on the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  parameters:
    SearchQueryParam:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: Keyword search query.
    CategoryFilterParam:
      name: categoryId
      in: query
      required: false
      schema:
        type: array # Allow multiple category selections
        items:
          type: string
          format: uuid
      style: form # How array parameters are formatted (e.g., categoryId=uuid1&categoryId=uuid2)
      explode: true
      description: Filter items by one or more category IDs.
    AuthorFilterParam:
      name: authorId
      in: query
      required: false
      schema:
        type: array
        items:
          type: string
          format: uuid
      style: form
      explode: true
      description: Filter items by one or more author IDs.
    PublisherFilterParam:
      name: publisherId
      in: query
      required: false
      schema:
        type: array
        items:
          type: string
          format: uuid
      style: form
      explode: true
      description: Filter items by one or more publisher IDs.
    PriceMinFilterParam:
      name: priceMin
      in: query
      required: false
      schema:
        type: number
        format: float
      description: Minimum price filter.
    PriceMaxFilterParam:
      name: priceMax
      in: query
      required: false
      schema:
        type: number
        format: float
      description: Maximum price filter.
    YearFilterParam:
      name: year
      in: query
      required: false
      schema:
        type: array # Allow filtering by specific years
        items:
          type: integer
      style: form
      explode: true
      description: Filter items by publication year(s).
    StockFilterParam:
      name: excludeOutOfStock
      in: query
      required: false
      schema:
        type: boolean
        default: false
      description: If true, exclude items that are out of stock.
    LanguageFilterParam:
      name: language
      in: query
      required: false
      schema:
        type: array # Allow filtering by language codes
        items:
          type: string
          enum: [TAMIL, ENGLISH]
      style: form
      explode: true
      description: Filter items by language code(s).
    ItemTypeFilterParam:
      name: itemType
      in: query
      required: false
      schema:
        type: array # Allow filtering by language codes
        items:
          type: string
          enum: [BOOK]
      style: form
      explode: true
      description: Filter items by language code(s).
    SortParam:
      name: sortBy
      in: query
      required: false
      schema:
        type: string
        enum: [relevance, best_sellers, price_asc, price_desc, name_asc, name_desc, year_newest, year_oldest]
        default: relevance
      description: Sorting criteria for the results.
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination.
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100 # Define a max page size
        default: 20
      description: Number of items per page.
  securitySchemes:
    bearerAuth: # Can be named anything, used later to reference
      type: http
      scheme: bearer
      bearerFormat: JWT # Optional, for documentation
      description: "JWT token obtained after successful OTP verification."